Move and 'purify' read_mappings
  This function doesn't need to have access to the mutex (or even need
  to know that it exists). The mutext is a concern of the API layer to
  ensure concurrent access to the state. Once acquired, it suffices to
  pass a mutable reference.

  This also fixes a small 'bug' in the `read_mappings` function which
  would log the *total* entries in map rather than the new inserted
  items.

  Note also that this function doesn't get rid of any removed token (or
  blacklisted ones).
