+++ src/server/api.rs
@@ -1,14 +1,13 @@
-use std::{
-    collections::HashMap,
-    fs::read_dir,
-    path::{Path, PathBuf},
-    sync::{Arc, Mutex},
-};
-
+use super::registry;
 use actix_web::{get, post, web, HttpResponse, Responder};
 use log;
 use serde::Deserialize;
 use serde_json::{json, Value};
+use std::{
+    collections::HashMap,
+    path::PathBuf,
+    sync::{Arc, Mutex},
+};
 
 pub struct AppMutState {
     pub mappings: Arc<Mutex<HashMap<String, serde_json::Value>>>,
@@ -25,44 +24,17 @@ pub async fn health() -> impl Responder {
 /// Endpoint to trigger update of the data
 #[get("/reread")]
 pub async fn reread_mappings(app_data: web::Data<AppMutState>) -> impl Responder {
-    read_mappings(&app_data.registry_path, &app_data.mappings);
-    HttpResponse::Ok().body("Reread contents")
-}
-
-/// Function that reads the files in registry_path and updates the mappings
-/// This function should add better error handling by returning a result so
-/// the views can act accordingly!
-pub fn read_mappings(registry_path: &Path, mappings: &Mutex<HashMap<String, serde_json::Value>>) {
-    log::debug!(
-        "Reading mappings from {}",
-        registry_path.to_str().unwrap_or("?")
-    );
-
-    // Can we create a ReadDir iterator of the PathBug?
-    let paths = read_dir(registry_path);
-
-    let mut mtx = match mappings.lock() {
+    let mut mtx = match app_data.mappings.lock() {
         Ok(mtx) => mtx,
         Err(e) => {
             log::error!("Error acquiring mutex lock! {}", e.to_string());
-            return;
+            return HttpResponse::InternalServerError().body("");
         }
     };
 
-    // We know paths is not an error
-    for path in paths.expect("Not a ReadDir Iterator") {
-        let path = path.expect("Not a DirEntry").path();
-        let stem_path = path.file_stem().expect("No file_stem");
-        let stem_str = stem_path.to_str().expect("Failed creating str");
-        // The key for hashmap is the name of the json file on disk
-        let key = stem_str.to_string();
-        if let Ok(raw_json) = std::fs::read_to_string(&path) {
-            if let Ok(json_data) = serde_json::from_str(&raw_json) {
-                mtx.insert(key, json_data);
-            }
-        }
-    }
-    log::info!("Read {} items", mtx.len());
+    registry::read_mappings(&app_data.registry_path, &mut mtx);
+
+    HttpResponse::Ok().body("Reread contents")
 }
 
 /// Endpoint to retrieve a single subject
